requirements.txt
pyyaml >= 6.0.2
beautifulsoup4 >= 4.13.4
requests >= 2.32.3
slixmpp >= 1.10.0
Mastodon.py >= 1.8.1
